SampleのyをDoublePlottableに限定した
