+++ app/src/main/java/com/nemo/coharts/chart/domain/plottable/Plottable.kt
@@ -1,0 +1,6 @@
+package com.nemo.coharts.chart.domain.plottable
+
+interface Plottable {
+    val label: String
+    val value: Any
+}
