--- app/src/main/java/com/nemo/coharts/chart/protocol/Sample.kt
@@ -1,6 +1,0 @@
-package com.nemo.coharts.chart.protocol
-
-data class Sample<X : Plottable, Y : Plottable>(
-    val xPlottable: X,
-    val yPlottable: Y
-)
